services/sensor/app/requirements.txt
numpy==1.21.0
pandas==1.3.0
python-dateutil==2.8.2
pytz==2021.1
kafka-python==2.1.3
pymongo==4.11.3